refactor: loading-animation.tsx dir
