+++ src/components/common/utilities/loading-animation.tsx
